% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_soilgrids.R
\encoding{UTF-8}
\name{add_soilgrids}
\alias{add_soilgrids}
\title{Adds SoilGrids soil description}
\usage{
add_soilgrids(
  x,
  soilgrids_path = NULL,
  widths = NULL,
  replace_existing = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{An object of class \code{\link{sf}} with a valid CRS definition. If it contains a column called 'land_cover_type', soils will be retrieved for "agriculture" and "wildland"
cover types. Otherwise, soils are retrieved for all locations.}

\item{soilgrids_path}{Path to SoilGrids rasters (see details). If missing, the SoilGrids REST API (https://rest.isric.org) will be queried.}

\item{widths}{A numeric vector indicating the desired layer widths, in \emph{mm}. If \code{NULL} the default soil grids layer definition is returned.}

\item{replace_existing}{A logical flag to force the replacement of existing soil data, when already present}

\item{verbose}{A logical flag to include a progress bar while processing the output of the query to the SoilGrids REST API.}
}
\value{
A modified object of class \code{\link{sf}} with column 'soil'.
}
\description{
Fills column 'soil' with information drawn from SoilGrids 2.0
}
\details{
If \code{soilgrids_path = NULL} the function connects with the SoilGrids REST API (https://rest.isric.org)
to retrieve the soil physical and chemical characteristics for a site (Hengl \emph{et al}. 2007), selected
by its coordinates. Also, in case the depths are not the default ones in the SoilGrids API, the function uses
averages the values of soil grid layers depending on the overlap between soil layer definitions.

If \code{soilgrids_path != NULL} the function will read SoilGrid rasters from the file disk. Folders need to be defined
for each variable ("sand", "clay", "soc", "bdod", "cfvo" and "nitrogen"). File paths from \code{soilgrids_path} should be named:

\emph{var}/\emph{var}_\emph{layer}_mean.tif

where \emph{var} is one of the above and \emph{layer} is "0-5cm", "5-15cm", "15-30cm", "30-60cm", "60-100cm" or "100-200cm"
}
\examples{
 \dontrun{
  # See vignette 'Landscape inputs'
 }

}
\references{
Hengl T, Mendes de Jesus J, Heuvelink GBM, Ruiperez Gonzalez M, Kilibarda M, \enc{Blagotić}{Blagotic} A, et al. (2017) SoilGrids250m: Global gridded soil information based on machine learning. PLoS ONE 12(2): e0169748. doi:10.1371/journal.pone.0169748.
}
\seealso{
\code{\link[=create_landscape]{create_landscape()}}, \code{\link[=impute_forests]{impute_forests()}}, \code{\link[medfate]{soil}}, \code{\link[medfate]{defaultSoilParams}}
}
\author{
\enc{Víctor}{Victor} Granda, EMF-CREAF

Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
