% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fire_regime.R
\name{create_fire_regime}
\alias{create_fire_regime}
\title{Create fire regime}
\usage{
create_fire_regime(annual_burned_area, doy = NULL, stochastic = FALSE)
}
\arguments{
\item{annual_burned_area}{A named vector of burned area in hectares for simulation years.}

\item{doy}{A named integer vector with the day of the year (i.e. between 1 and 366) when fires will be simulated for each simulation year in \code{annual_burned_area}. If NULL
fires will be simulated on the driest day (i.e. when vapor pressure deficit is largest).}

\item{stochastic}{A logical flag to allow for stochastic fire simulation. If FALSE, the values in \code{annual_burned_area} are
taken directly as target area to be burned. If TRUE, the target area to be burned is assumed to follow
a Poisson distribution of lambda equal to the supplied value.}
}
\value{
A list with the supplied parameters
}
\description{
Defines an object containing fire regime parameters for simulations of forest dynamics.
}
\examples{
freg <- create_fire_regime(c("2002" = 1000, "2003" = 5000)) 

}
\seealso{
\code{\link{fordyn_scenario}}, \code{\link{fordyn_spatial}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
