% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_landscape.R
\name{create_landscape}
\alias{create_landscape}
\title{Creates a landscape}
\usage{
create_landscape(
  x,
  dem,
  land_cover_map = NULL,
  wildland = NULL,
  agriculture = NULL,
  rock = NULL,
  artificial = NULL,
  water = NULL,
  progress = TRUE
)
}
\arguments{
\item{x}{An object of class \code{\link{sf}}}

\item{dem}{A digital elevation model (class \code{\link{SpatRaster}}) with meters as units}

\item{land_cover_map}{An object of class \code{\link{SpatRaster}} of land cover type. If missing, all locations are considered 'wildland'.}

\item{wildland, agriculture, rock, artificial, water}{Strings indicating the mapping from the legend of land_cover_map.}

\item{progress}{A logical flag to print console output}
}
\value{
A modified object of class \code{\link{sf}} with columns:
\itemize{
\item{\code{elevation}: Elevation above sea level (in m).}
\item{\code{slope}: Slope (in degrees).}
\item{\code{aspect}: Aspect (in degrees).}
\item{\code{land_cover_type}: Land cover type.}
}
}
\description{
Initializes topography and land cover type for a set of target locations
}
\details{
The user should manually define the mapping of land cover classes in \code{land_cover_map} to the land cover types
used in medfateland.
}
\examples{
\dontrun{
  # See vignette 'Preparing inputs'
}
}
\seealso{
\code{\link[=impute_forests]{impute_forests()}}, \code{\link[=add_soilgrids]{add_soilgrids()}}
}
