% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_management_scenario.R
\name{create_management_scenario}
\alias{create_management_scenario}
\title{Create management scenario}
\usage{
create_management_scenario(
  numberOfUnits,
  annualDemandBySpecies = numeric(0),
  managementFunction = NULL,
  managementArguments = NULL
)
}
\arguments{
\item{numberOfUnits}{Number of management units. Each management unit represents a group of forest stands following the same silvicultural model.}

\item{annualDemandBySpecies}{A vector of annual wood demand (m3) by medfate species names. If empty, the scenario is not demand-based.}

\item{managementFunction}{A function that implements forest management actions (see \code{\link{fordyn}}). If NULL, function \code{\link{defaultManagementFunction}} will be taken.}

\item{managementArguments}{A list of arguments to be passed to the managementFunction. These arguments will be taken as defaults to be copied for all management units and can later be modified.
If NULL, the result of calling function \code{\link{defaultManagementArguments}} will be taken.}
}
\value{
A list with the following structure:
\itemize{
\item{\code{managementFunction}: The supplied management function}
\item{\code{managementUnits}: A vector of length numberOfUnits, each with the supplied management arguments}
\item{\code{annualDemandBySpecies}: A vector of annual wood demand (m3) by medfate species names}
}
}
\description{
Defines a management scenario with default values, to be modified by the user
}
\examples{

# A scenario with three management units and demand for two species
s = create_management_scenario(3,  c("Quercus ilex" = 1000, "Pinus nigra" = 2000))

}
\seealso{
\code{\link{fordyn_scenario}}
}
