% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_dispersal_control.R
\name{default_dispersal_control}
\alias{default_dispersal_control}
\title{Default control parameters for dispersal}
\usage{
default_dispersal_control()
}
\value{
A list with the following items:
\itemize{
\item{\code{distance_step  [= 25]}: Distance step in meters.}
\item{\code{maximum_dispersal_distance [= 3000]}: Maximum dispersal distance in meters.}
\item{\code{min_percent [= 1]}: A minimum percent of seed bank to retain entry in \code{seedBank} element of \code{forest}.}
\item{\code{stochastic_resampling [= FALSE]}: A flag to indicate that stochastic resampling of stands is performed.}
}
}
\description{
Defines default control parameters for dispersal process
}
\examples{
default_dispersal_control()

}
\seealso{
\code{\link{spwb_land}}, \code{\link{fordyn_scenario}} \code{\link{dispersal}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
