% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_watershed_control.R
\name{default_watershed_control}
\alias{default_watershed_control}
\title{Default control parameters for watershed processes}
\usage{
default_watershed_control(watershed_model = "tetis")
}
\arguments{
\item{watershed_model}{Either "tetis" or "serghei"}
}
\value{
A list with the following items:
\itemize{
\item{\code{watershed_model}: A string with the watershed model.}
\item{\code{tetis_correction_factors}: A list with the following elements:
\itemize{
\item{\code{Rdrain}: Correction factor for vertical hydraulic saturated conductivity between soil and aquifer.}
\item{\code{Rinterflow}: Correction factor for soil hydraulic saturated conductivity (subsurface flow between grid cells).}
\item{\code{Rbaseflow}: Correction factor for bedrock hydraulic conductivity (groundwaterflow between grid cells).}
}
}
\item{\code{serghei_input_dir}: Path to SERGHEI input files.}
\item{\code{serghei_output_dir}: Path to SERGHEI output files.}
}
}
\description{
Defines default control parameters for watershed processes
}
\examples{
default_watershed_control()

}
\seealso{
\code{\link{spwb_land}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
