% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{defaultPrescriptionsBySpecies}
\alias{defaultPrescriptionsBySpecies}
\title{Default prescriptions by species}
\source{
Mario Beltrán & Míriam Piqué. Forest Science and Technology Centre of Catalonia (CTFC)
}
\description{
Default management prescriptions by species, defined according current practices in Catalonia (NE Spain)
}
\details{
A data frame with 27 species (or species groups) in rows and management parameters in columns (defined in \code{\link[medfate]{defaultManagementArguments}})
}
\seealso{
\code{\link{create_management_scenario}}, \code{\link[medfate]{defaultManagementArguments}},
\code{\link{fordyn_scenario}}
}
\keyword{data}
