% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultWatershedCorrectionFactors.R
\name{defaultWatershedCorrectionFactors}
\alias{defaultWatershedCorrectionFactors}
\title{Default correction factors for watershed hydraulics}
\usage{
defaultWatershedCorrectionFactors()
}
\value{
A list with the following items:
\itemize{
\item{\code{Rdrain}: Correction factor for vertical hydraulic saturated conductivity between soil and aquifer.}
\item{\code{Rinterflow}: Correction factor for soil hydraulic saturated conductivity (subsurface flow between grid cells).}
\item{\code{Rbaseflow}: Correction factor for bedrock hydraulic conductivity (groundwaterflow between grid cells).}
}
}
\description{
Defines default values for correction factors used to modulate watershed hydraulic parameters
}
\examples{
defaultWatershedCorrectionFactors()
}
\seealso{
\code{\link{spwb_land}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
