% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{example_ifn}
\alias{example_ifn}
\title{Example of distributed forest inventory stands}
\format{
The data format is that of an object \code{\link[sf]{sf}}
}
\source{
\itemize{
\item{Soil data from SoilGrids global database (Hengl et al. 2017).}
\item{Soil depth and depth to bedrock from Shangguan et al. (2017).}
\item{Forest structure and composition from the Third Spanish Forest Inventory (IFN3).}
}
}
\description{
An example of an coordinates, topography, forest and soil data corresponding to 30 forest inventory plots.
}
\seealso{
{\code{\link{spwb_spatial}}}
}
