% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_regime_instance.R
\name{fire_regime_instance}
\alias{fire_regime_instance}
\title{Fire regime instance}
\usage{
fire_regime_instance(sf, fire_regime)
}
\arguments{
\item{sf}{An object of class \code{\link{sf}} with the following columns:
\itemize{
\item{\code{geometry}: Spatial geometry.}
\item{\code{id}: Stand identifiers.}
\item{\code{represented_area}: Area represented by each stand (in hectares).}
\item{\code{ignition_weights}: Relative weights to determine stands to be burned (optional).}
}}

\item{fire_regime}{A list of parameters defining the fire regime (see \code{\link{create_fire_regime}}).}
}
\value{
An integer matrix specifying the day of the year of burning of each landscape unit for every year in the fire regime definition
}
\description{
Applies a fire regime object over a set of landscape units to determine a fire realization
}
\seealso{
\code{\link{create_fire_regime}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
