% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fordynscenario.R
\name{fordynscenario}
\alias{fordynscenario}
\title{Scenario of forest dynamics}
\usage{
fordynscenario(
  y,
  SpParams,
  meteo,
  localControl = defaultControl(),
  dates = NULL,
  managementScenario = NULL,
  CO2ByYear = numeric(0),
  parallelize = FALSE,
  numCores = detectCores() - 1,
  chunk.size = NULL,
  progress = TRUE
)
}
\arguments{
\item{y}{An object of class \code{\link{SpatialPointsLandscape-class}}, \code{\link{SpatialPixelsLandscape-class}} or \code{\link{SpatialGridLandscape-class}}.
Management units should be defined in this object.}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{meteo}{Meteorology data (see \code{\link{fordynspatial}}).}

\item{localControl}{A list of local model control parameters (see \code{\link{defaultControl}}).}

\item{dates}{A \code{\link{Date}} object with the days of the period to be simulated. If \code{NULL}, then the whole period of \code{meteo} is used.}

\item{managementScenario}{A list defining the management scenario (i.e. year-by-year wood demand, management arguments and management function by management unit)}

\item{CO2ByYear}{A named numeric vector with years as names and atmospheric CO2 concentration (in ppm) as values. Used to specify annual changes in CO2 concentration along the simulation (as an alternative to specifying daily values in \code{meteo}).}

\item{parallelize}{Boolean flag to try parallelization (will use all clusters minus one).}

\item{numCores}{Integer with the number of cores to be used for parallel computation.}

\item{chunk.size}{Integer indicating the size of chuncks to be sent to different processes (by default, the number of spatial elements divided by the number of cores).}

\item{progress}{Boolean flag to display progress information for simulations.}
}
\description{
Scenario of forest dynamics
}
