% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_forests.R
\name{impute_forests}
\alias{impute_forests}
\title{Forest imputation}
\usage{
impute_forests(
  x,
  sf_nfi,
  dem,
  forest_map,
  varClass,
  replace_existing = FALSE,
  empty_forests = TRUE,
  merge_trees = TRUE,
  merge_shrubs = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{An object of class \code{\link{sf}}}

\item{sf_nfi}{An object of class \code{\link{sf}} with forest inventory data column 'forest'}

\item{dem}{A digital elevation model (class \code{\link{rast}}) with meters as units}

\item{forest_map}{An object of class \code{\link{rast}} or \code{\link{vect}} with the forest class map}

\item{varClass}{Variable name or index containing forest classes in 'forest_map'}

\item{replace_existing}{A logical flag to force the replacement of existing 'forest' objects, when present}

\item{empty_forests}{A logical flag to place empty 'forest' objects for wildland cells not having a forest class}

\item{merge_trees}{A logical flag to simplify tree cohorts by merging tree records in DBH classes (see \code{\link{forest_mergeTrees}})}

\item{merge_shrubs}{A logical flag to simplify shrub cohorts by merging shrub records in height classes (see \code{\link{forest_mergeShrubs}})}

\item{verbose}{A logical flag to print console output.}
}
\value{
A modified object of class \code{\link{sf}} with column 'forest'.
}
\description{
Performs imputation of forest objects from a forest inventory using a forest class map to match composition and topography as covariates.
}
\examples{
\dontrun{
  # See vignette 'Landscape inputs'
}
}
