% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape_summary.R
\name{landscape_summary}
\alias{landscape_summary}
\title{Forest and soil summaries over space}
\usage{
landscape_summary(object, name, summaryFunction, ...)
}
\arguments{
\item{object}{An object of class \code{\link{sf}}.}

\item{name}{A string of the element to summarize: "forest", "soil" or "state".}

\item{summaryFunction}{A function that accepts objects of class \code{\link{forest}}, \code{soil} or model input objects, respectively.}

\item{...}{Additional arguments to the summary function.}
}
\value{
An object of class \code{\link{sf}} containing the calculated statistics.
}
\description{
Functions to calculates a summary function for the forest or soil of all spatial elements 
in an object of class \code{\link{sf}} containing landscape information.
}
\examples{
# Load plot data and species parameters from medfate
data(examplepointslandscape)

# Load default medfate parameters
data("SpParamsMED")
 
# Transform example to 'sf' 
y = sp_to_sf(examplepointslandscape)

# Apply forest summary function
landscape_summary(y, "forest", summary.forest, SpParamsMED)
 
}
\seealso{
\code{\link{forest}}, \code{\link{soil}}, \code{\link{summary.forest}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
}
