% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_land.R
\name{overland_routing}
\alias{overland_routing}
\title{Overland routing for TETIS sub-model}
\usage{
overland_routing(r, sf)
}
\arguments{
\item{r}{An object of class \code{\link[terra]{SpatRaster}}, defining the raster topology.}

\item{sf}{An object of class \code{\link[sf]{sf}} with the following columns:
\itemize{
\item{\code{geometry}: Spatial point geometry corresponding to cell centers.}
\item{\code{elevation}: Elevation above sea level (in m).}
\item{\code{channel}: An optional logical (or binary) vector indicating overland channel routing.}
}}
}
\value{
An object of class \code{\link[sf]{sf}} describing overland routing parameters and outlet cells:
\itemize{
\item{\code{geometry}: Spatial point geometry corresponding to cell centers.}
\item{\code{elevation}: Elevation above sea level (in m).}
\item{\code{waterRank}: Ranked elevation in decreasing order.}
\item{\code{waterOrder}: A vector with the cell's processing order for overland routing (based on elevation). First value corresponds to the row index of the first processed cell, second value corresponds to the row index of the second processed cell and so forth.}
\item{\code{queenNeigh}: A list where, for each cell, a vector gives the identity of neighbours (up to eight).}
\item{\code{waterQ}: A list where, for each cell, a vector gives the proportion of overland flow to each neighbour.}
\item{\code{outlet}: A logical vector indicating outlet/channel cells.}
}
}
\description{
Determines overland routing given a raster definition and a set of target locations for watershed simulations.
}
\details{
If \code{channel} is supplied, then all those cells are considered outlet for the sake of runoff routing.
}
\examples{
# Load example watershed data
data("example_watershed")

# Get bounding box to determine limits
b <- sf::st_bbox(example_watershed)
b

# Define a raster topology, using terra package, 
# with the same CRS as the watershed. In this example cells have 100 m side.
# Coordinates in the 'sf' object are assumed to be cell centers
r <-terra::rast(xmin = 401380, ymin = 4671820, xmax = 402880, ymax = 4672620, 
                nrow = 8, ncol = 15, crs = "epsg:32631")
                
# Generate overland routing
or <- overland_routing(r, example_watershed)

# Plot elevation
plot(or["elevation"])

# Rank (decreasing elevation) for processing
plot(or["waterRank"])

# Plot outlet cells
plot(or["outlet"])
}
