% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DefineSpatialClasses.R
\docType{class}
\name{SFLandscape-class}
\alias{SFLandscape-class}
\title{Class \code{"SFLandscape"}}
\arguments{
\item{sf}{Object of class "sf" containing coordinates, topography, etc of every simple feature.}

\item{forestlist}{Object of class \code{"list"} containing a set of \code{\link{forest}} objects.}

\item{soillist}{Object of class \code{"list"} containing a set of \code{\link{soil}} objects.}

\item{xlist}{Object of class \code{"list"} containing a set of spwb or growth model input objects.}
}
\description{
An S4 class that represents a landscape configuration over a set of simple features.
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
