% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplot_land.R
\name{shinyplot.spwb_land}
\alias{shinyplot.spwb_land}
\alias{shinyplot.growth_land}
\alias{shinyplot.fordyn_land}
\title{Shiny app with interactive plots and maps of watershed simulation results}
\usage{
\method{shinyplot}{spwb_land}(x, r = NULL, ...)

\method{shinyplot}{growth_land}(x, r = NULL, ...)

\method{shinyplot}{fordyn_land}(x, r = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{spwb_land}}, \code{\link{growth_land}} or \code{\link{fordyn_land}}.}

\item{r}{An object of class \code{\link[terra]{SpatRaster}}, defining the raster topology.}

\item{...}{Additional parameters for function shinyplot (not used).}
}
\value{
An object that represents the shiny app
}
\description{
Creates a shiny app with interactive plots for simulation results
}
\seealso{
\code{\link{spwb_land}}, \code{\link{plot.spwb_land}}, \code{\link{plot_variable}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
