% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_to_sf.R
\name{sp_to_sf}
\alias{sp_to_sf}
\title{sp to sf conversion}
\usage{
sp_to_sf(y)
}
\arguments{
\item{y}{An object of class 'Spatial_*_Landscape' or 'DistributedWatershed'}
}
\value{
An object of class 'sf'
}
\description{
Converts objects of older classes to sf
}
