% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineConstructors.R
\name{SpatialPointsLandscape}
\alias{SpatialPointsLandscape}
\alias{SpatialGridLandscape}
\alias{SpatialPixelsLandscape}
\title{Creates spatial landscape objects}
\usage{
SpatialPointsLandscape(spt, lct, forestlist, soillist)

SpatialGridLandscape(sgt, lct, forestlist, soillist)

SpatialPixelsLandscape(spxt, lct, forestlist, soillist)
}
\arguments{
\item{spt}{An object of class \code{\link{SpatialPointsTopography}}.}

\item{lct}{A named character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock' or 'static').}

\item{forestlist}{A list of objects of class 'forest' with the same number of elements as spatial points/pixels.}

\item{soillist}{A list of objects of class 'forest' with the same number of elements as spatial points/pixels.}

\item{sgt}{An object of class \code{\link{SpatialGridTopography}}.}

\item{spxt}{An object of class \code{\link{SpatialPixelsTopography}}.}
}
\value{
An object of class \code{\link{SpatialPointsLandscape-class}}, \code{\link{SpatialPixelsLandscape-class}} or \code{\link{SpatialGridLandscape-class}} depending on the function.
}
\description{
Functions to instantiate spatial landscape classes.
}
\details{
Function \code{SpatialPointsLandscape()} requires \emph{named} vectors/lists for parameters \code{forestlist}, 
\code{soillist} and \code{lct}, whereas row names in the coordinate slot of \code{spt} also needs to be defined. 
This requirement ensures that if the number of elements in \code{spt} or these lists differs only the common points 
identified by name will be part of the returned object (a warning is launched in these cases). 
Name matching is not performed for \code{SpatialGridLandscape()} and \code{SpatialPixelsLandscape()},
which require the same number of grid cells as vector elements and assume that these are supplied in the right order.
}
\seealso{
\code{\link{forest}}, \code{\link{soil}}, \code{\link{DistributedWatershed}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
