% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spwbspatial.R
\name{summary.spwbspatial}
\alias{summary.spwbspatial}
\alias{summary.growthspatial}
\alias{summary.fordynspatial}
\title{Summarizes spatial simulation results}
\usage{
summary.spwbspatial(object, summaryFunction, ...)

summary.growthspatial(object, summaryFunction, ...)

summary.fordynspatial(object, summaryFunction, ...)
}
\arguments{
\item{object}{An object of the appropriate class, containing simulation results.}

\item{summaryFunction}{The summary function to be executed on simulation results (see details).}

\item{...}{Additional parameters to the summary function.}
}
\value{
An object of class \code{\link{summaryspatial}}, with the following two elements:
 \itemize{
   \item{\code{sp}: An object with spatial information (of \code{SpatialPoints-class}, \code{SpatialPixels-class} or \code{SpatialGrid-class}).}
   \item{\code{summarylist}: A list of model output summaries for each simulated stand.}
 }
}
\description{
Creates spatial objects containing summaries of simulations
}
\details{
The function supplied should take as input an object of local simulation function, i.e. \code{\link{spwb}}, \code{\link{growth}}, or \code{\link{fordyn}}. 
The output should be a matrix with dates as rows and variables in columns. 
An example of suitable function is \code{\link{summary.spwb}}.
}
\seealso{
\code{\link{spwbspatial}}, \code{\link{plot.summaryspatial}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
}
