% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{unnest_summary}
\alias{unnest_summary}
\title{Extracts cell simulation summaries}
\usage{
unnest_summary(x)
}
\arguments{
\item{x}{An object of class \code{\link[sf]{sf}}, with simulation summaries. Alternatively an object of class  \code{\link{spwb_land}}, \code{\link{growth_land}} or \code{\link{fordyn_land}}.}
}
\value{
An sf object with all cell summaries.
}
\description{
Extracts cell simulation summaries
}
\details{
Appropriate input objects for \code{x} can be originated from calls to \code{\link{simulation_summary}}.
Alternatively, if summary functions were specified at the time of performing simulations,
the result of the spatial simulation function (e.g. \code{\link{spwb_spatial}})
will already contain the summaries. A special case is made for \code{\link{spwb_land}}, \code{\link{growth_land}},
and \code{\link{fordyn_land}}, that are accepted inputs as \code{x}, because their child element 'sf' is used.
}
\seealso{
\code{\link{plot_summary}}, \code{\link{spwb_spatial}}, \code{\link{spwb_land}},  \code{\link{simulation_summary}}
}
