% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_state.R
\name{update_state}
\alias{update_state}
\title{Updates the state of a landscape object}
\usage{
update_state(x, y)
}
\arguments{
\item{x}{An object of class \code{\link{sf}} with the corresponding landscape columns.}

\item{y}{The object resulting of a simulation previously carried on \code{x}.}
}
\value{
An object of class \code{\link{sf}} with modified state variables.
}
\description{
Updates the state of a spatial object 'x' according to the final state in simulation outcome 'y'
}
\seealso{
\code{\link{spwb_spatial}}, \code{\link{spwb_spatial_day}}, \code{\link{spwb_land}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
}
