% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateState.R
\name{updateState}
\alias{updateState}
\title{Updates the state of a landscape object}
\usage{
updateState(x, y)
}
\arguments{
\item{y}{The object resulting of a simulation previously carried on \code{x}.}
}
\value{
An object of class \code{\link{SpatialPointsLandscape-class}}, \code{\link{SpatialPixelsLandscape-class}} or \code{\link{SpatialGridLandscape-class}} with modified state variables.
}
\description{
Updates the state of a spatial object 'x' according to the final state in simulation outcome 'y'
}
\seealso{
\code{\link{spwbspatial}}, \code{\link{spwbspatial_day}}, \code{\link{spwbland}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
}
