% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_watershed_control.R
\name{default_watershed_control}
\alias{default_watershed_control}
\title{Default control parameters for watershed processes}
\usage{
default_watershed_control(watershed_model = "tetis")
}
\arguments{
\item{watershed_model}{Hydrological model for watershed processes. Only "tetis" or "serghei" are accepted.}
}
\value{
A list with the following items:
\itemize{
\item{\code{watershed_model}: A string with the watershed model.}
\item{\code{weather_aggregation_factor [= 1]}: An integer specifying the spatial aggregation for interpolated weather.}
\item{\code{tetis_parameters}: A list of TETIS parameters with the following elements:
\itemize{
\item{\code{R_localflow [= 1.0]}: Correction factor for soil hydraulic saturated conductivity (local vertical flows).}
\item{\code{R_interflow [= 100.0]}: Correction factor for soil hydraulic saturated conductivity (subsurface flow between grid cells).}
\item{\code{R_baseflow [= 1.0]}: Correction factor for bedrock hydraulic conductivity (groundwaterflow between grid cells).}
\item{\code{n_interflow [= 1.0]}: Exponent for the determination of interflow.}
\item{\code{n_baseflow [= 1.0]}: Exponent for the determination of baseflow.}
\item{\code{num_daily_substeps [= 4]}: Number of daily sub-steps for interflow calculations.}
\item{\code{rock_max_infiltration [= 10]}: Maximum infiltration rate (mm·day-1) for rock cells.}
\item{\code{deep_aquifer_loss [= 0]}: Daily loss rate from watershed aquifer towards a deeper aquifer not connected to outlets (mm·day-1).}
}
}
\item{\code{serghei_parameters}: A list of SERGHEI parameters with the following elements:
\itemize{
\item{\code{input_dir [= ""]}: Path to SERGHEI input files.}
\item{\code{output_dir [= ""]}: Path to SERGHEI output files.}
\item{\code{force_equal_layer_widths [= FALSE]}: A boolean flag to force equal layer widths (taken from the first soil element) in all soils.}
}
}
\item{\code{dispersal_parameters}: A list of dispersal parameters (only for \code{\link{fordyn_land}}) with the following elements:
\itemize{
\item{\code{distance_step  [= 25]}: Distance step in meters.}
\item{\code{maximum_dispersal_distance [= 3000]}: Maximum dispersal distance in meters.}
\item{\code{min_percent [= 1]}: A minimum percent of seed bank to retain entry in \code{seedBank} element of \code{forest}.}
\item{\code{stochastic_resampling [= FALSE]}: A flag to indicate that stochastic resampling of stands is performed.}
}
}
}
}
\description{
Defines default control parameters for watershed processes
}
\examples{
default_watershed_control()

}
\seealso{
\code{\link{spwb_land}},  \code{\link{dispersal}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
