% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_utils.R
\name{dispersal}
\alias{dispersal}
\title{Seed production and dispersal}
\usage{
dispersal(
  sf,
  SpParams,
  local_control = medfate::defaultControl(),
  order = 4,
  minPercent = 1
)
}
\arguments{
\item{sf}{An object of class \code{\link{sf}} with the following columns:
\itemize{
\item{\code{geometry}: Spatial geometry.}
\item{\code{land_cover_type}: Land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').}
\item{\code{forest}: Objects of class \code{\link{forest}}.}
\item{\code{queenNeigh}: A list of integers identifying the (up to 8) queen neighbors, for each cell.}
\item{\code{represented_area}: Area represented by each cell (in m2).}
}}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{local_control}{A list of control parameters (see \code{\link{defaultControl}})}

\item{order}{Order of neighbors in a grid}

\item{minPercent}{A minimum percent of seed bank to retain entry in \code{seedBank} element of \code{forest}.}
}
\value{
A list with forest objects (for wildland cover type) containing a modified seed bank
}
\description{
Simulates seed bank mortality, seed production and dispersal among grid cells in a watershed
}
\details{
Dispersal kernel follows Clark et al. (1999) and potential seed donors (neighbors) are defined up to a given grid distance order.
The sum of dispersal kernel values over the neighbors plus the kernel of the local
grid cell is used to define the normalizing constant for the kernel. A maximum value of 100\\% seed bank refill is attained for
species with seed production in all seed donors and the local cell.
}
\examples{
data(example_watershed)
data(SpParamsMED)

forest_list <- dispersal(example_watershed, SpParamsMED)

forest_list[[1]]$seedBank
}
\references{
Clark et al. (1999). Seed dispersal near and far: Patterns across temperate and tropical forests. Ecology 80(5):1475-1494
}
\seealso{
\code{\link{fordyn_land}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF.
}
