% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DefineSpatialClasses.R
\docType{class}
\name{DistributedWatershed-class}
\alias{DistributedWatershed-class}
\title{Class \code{"DistributedWatershed"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class that represents a landscape configuration over a grid of coordinates.
}
\section{Slots}{

\describe{
\item{\code{coords}}{A matrix of pixel coordinates.}

\item{\code{coords.nrs}}{Inherited from \code{SpatialPointsDataFrame} but not used.}

\item{\code{grid}}{Object of class \code{"GridTopology"}.}

\item{\code{grid.index}}{Index of points in full grid.}

\item{\code{bbox}}{Object of class \code{"matrix"} with the boundary box.}

\item{\code{proj4string}}{Object of class \code{"CRS"} with the projection string.}

\item{\code{data}}{Object of class "data.frame" containing the elevation (in m), slope (in degrees) and aspect (in degrees from North) of every cell.}

\item{\code{lct}}{A character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').}

\item{\code{waterOrder}}{A numeric vector of cell processing order.}

\item{\code{waterQ}}{A list of water discharge values to neighbours.}

\item{\code{queenNeigh}}{A list of neighbours for each cell.}

\item{\code{channel}}{A logical vector indicating whether each cell belongs to the channel network.}

\item{\code{bedrock}}{A data frame of bedrock properties (depth to bedrock, conductivity and porosity).}

\item{\code{aquifer}}{A numeric vector with the water content of the aquifer in each cell.}

\item{\code{snowpack}}{A numeric vector with the snow water equivalent content of the snowpack in each cell.}

\item{\code{forestlist}}{Object of class \code{"list"} containing a set of \code{\link{forest}} objects.}

\item{\code{soillist}}{Object of class \code{"list"} containing a set of \code{\link{soil}} objects.}

\item{\code{xlist}}{Object of class \code{"list"} containing a set of spwb or growth model input objects.}
}}

\examples{
#Structure of the S4 object
showClass("DistributedWatershed")

}
\seealso{
\code{\link{SpatialPixelsTopography-class}}, \code{\link{SpatialPixelsLandscape-class}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
