% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_regime_instance.R
\name{fire_regime_instance}
\alias{fire_regime_instance}
\title{Fire regime instance}
\usage{
fire_regime_instance(sf, fire_regime)
}
\arguments{
\item{sf}{An object of class \code{\link{sf}} with the following columns:
\itemize{
\item{\code{geometry}: Spatial geometry.}
\item{\code{id}: Stand identifiers.}
\item{\code{represented_area}: Area represented by each stand (in hectares).}
\item{\code{ignition_weights}: Relative weights to determine stands to be burned (optional).}
}}

\item{fire_regime}{A list of parameters defining the fire regime (see \code{\link{create_fire_regime}}).}
}
\value{
An integer matrix specifying the day of the year of burning of each landscape unit for every year in the fire regime definition.
Values are interpreted as follows:
\itemize{
\item{NA - No wildfire this year}
\item{0 - Wildfire will occur the driest day (i.e. the one with largest vapor pressure deficit).}
\item{1...366 - Day of the year when wildfire will occur}
}
}
\description{
Applies a fire regime object over a set of landscape units to determine a fire realization
}
\details{
The function randomly determines the landscape units that will burn every year, depending on the specifications
of the fire regime object.
}
\examples{
# Load example data
data("example_ifn")

# Assume that each stand represents 1km2 = 100 ha
example_ifn$represented_area <- 100

# Define fire regime characteristics
freg <- create_fire_regime(c("2002" = 200, "2003" = 500)) 

# Create a fire regime instance
fire_regime_instance(example_ifn, freg)

}
\seealso{
\code{\link{create_fire_regime}}, \code{\link{fordyn_spatial}}, \code{\link{fordyn_scenario}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
