% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fordyn_scenario.R
\name{fordyn_scenario}
\alias{fordyn_scenario}
\title{Scenario of forest dynamics}
\usage{
fordyn_scenario(
  y,
  SpParams,
  meteo,
  managementScenario,
  volumeFunction = NULL,
  localControl = defaultControl(),
  dates = NULL,
  CO2ByYear = numeric(0),
  summaryFunction = NULL,
  summaryArgs = NULL,
  parallelize = FALSE,
  numCores = detectCores() - 1,
  chunk.size = NULL,
  progress = TRUE
)
}
\arguments{
\item{y}{An object of class \code{\link{sf}} with landscape information.
Management units should be defined in this object.}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{meteo}{Meteorology data (see \code{\link{fordyn_spatial}}).}

\item{managementScenario}{A list defining the management scenario (see \code{\link{create_management_scenario}})}

\item{volumeFunction}{A function accepting a forest object as input and returning the wood volume (m3/ha) corresponding to each tree cohort.
If NULL, the default volume function is used (not recommended!).}

\item{localControl}{A list of local model control parameters (see \code{\link{defaultControl}}).}

\item{dates}{A \code{\link{Date}} object with the days of the period to be simulated. If \code{NULL}, then the whole period of \code{meteo} is used.}

\item{CO2ByYear}{A named numeric vector with years as names and atmospheric CO2 concentration (in ppm) as values. Used to specify annual changes in CO2 concentration along the simulation (as an alternative to specifying daily values in \code{meteo}).}

\item{summaryFunction}{An appropriate function to calculate summaries from an object of class 'fordyn' (e.g., \code{\link{summary.fordyn}}).}

\item{summaryArgs}{List with additional arguments for the summary function.}

\item{parallelize}{Boolean flag to try parallelization (will use all clusters minus one).}

\item{numCores}{Integer with the number of cores to be used for parallel computation.}

\item{chunk.size}{Integer indicating the size of chunks to be sent to different processes (by default, the number of spatial elements divided by the number of cores).}

\item{progress}{Boolean flag to display progress information for simulations.}
}
\value{
An list of class 'fordyn_scenario' with the following elements:
\itemize{
\item{An object of class 'sf' containing four elements:
\itemize{
\item{\code{geometry}: Spatial geometry.}
\item{\code{id}: Stand id, taken from the input.}
\item{\code{state}: A list of \code{\link{spwbInput}} or \code{\link{growthInput}} objects for each simulated stand, to be used in subsequent simulations (see \code{\link{update_landscape}}).}
\item{\code{forest}: A list of \code{\link{forest}} objects for each simulated stand (only in \code{fordynspatial}), to be used in subsequent simulations (see \code{\link{update_landscape}}).}
\item{\code{managementarguments}: A list of management arguments for each simulated stand (only in \code{fordynspatial}), to be used in subsequent simulations (see \code{\link{update_landscape}}).}
\item{\code{treetable}: A list of data frames for each simulated stand, containing the living trees at each time step.}
\item{\code{shrubtable}: A list of data frames for each simulated stand, containing the living shrub at each time step.}
\item{\code{deadtreetable}: A list of data frames for each simulated stand, containing the dead trees at each time step.}
\item{\code{deadshrubtable}: A list of data frames for each simulated stand, containing the dead shrub at each time step.}
\item{\code{cuttreetable}: A list of data frames for each simulated stand, containing the cut trees at each time step.}
\item{\code{cutshrubtable}: A list of data frames for each simulated stand, containing the cut shrub at each time step.}
\item{\code{summary}: A list of model output summaries for each simulated stand (if \code{summaryFunction} was not \code{NULL}).}
}
}
}
}
\description{
Evaluates forest dynamics over a landscape including climate and management scenarios
}
\examples{
\dontrun{
# Load example landscape data
data("examplepointslandscape")
  
# Transform example to 'sf' 
y = sp_to_sf(examplepointslandscape)

# Load example meteo data frame from package meteoland
data("examplemeteo")
  
#Prepare a two-year meteorological data with half precipitation during 
#the second year
meteo2001 = examplemeteo
meteo2002 = examplemeteo
meteo2002$Precipitation = meteo2002$Precipitation/2
row.names(meteo2002) = seq(as.Date("2002-01-01"), 
                         as.Date("2002-12-31"), by="day")
meteo_01_02 = rbind(meteo2001, meteo2002)
                         
# Load default medfate parameters
data("SpParamsMED")

# Creates default scenario with two management units and demand by species
s = create_management_scenario(2,  c("Quercus ilex" = 1000, "Pinus nigra" = 2000))

# Modify thinning threshold of the arguments of management unit #1
s$managementUnits[[1]]$managementArgs$thinningThreshold = 15

# Subset 10 stands for computational simplicity
y = y[1:10, ]

# Assign two management units to stands
y$managementunit[1:3] = 1 
y$managementunit[4:10] = 2

# Assume that each stand represents 1km2 = 100 ha
y$representedarea = 100

# Launch simulation scenario
res = fordyn_scenario(y, SpParamsMED, meteo_01_02, 
                      volumeFunction = NULL, managementScenario = s,
                      parallelize = TRUE)
}

}
\seealso{
\code{\link{fordyn_spatial}}, \code{\link{create_management_scenario}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
