% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fordyn_scenario.R
\name{fordyn_scenario}
\alias{fordyn_scenario}
\title{Scenario of forest dynamics}
\usage{
fordyn_scenario(
  sf,
  SpParams,
  meteo = NULL,
  management_scenario,
  volume_function = NULL,
  local_control = defaultControl(),
  dates = NULL,
  CO2ByYear = numeric(0),
  summary_function = NULL,
  summary_arguments = NULL,
  parallelize = FALSE,
  num_cores = detectCores() - 1,
  chunk_size = NULL,
  progress = TRUE
)
}
\arguments{
\item{sf}{An object of class \code{\link{sf}} with the following columns:
\itemize{
\item{\code{geometry}: Spatial geometry.}
\item{\code{id}: Stand identifiers.}
\item{\code{elevation}: Elevation above sea level (in m).}
\item{\code{slope}: Slope (in degrees).}
\item{\code{aspect}: Aspect (in degrees).}
\item{\code{forest}: Objects of class \code{\link{forest}}.}
\item{\code{soil}: Objects of class \code{\link{soil}}.}
\item{\code{state}: Objects of class \code{\link{spwbInput}} or \code{\link{growthInput}} (optional).}
\item{\code{meteo}: Data frames with weather data (required if parameter \code{meteo = NULL}).}
\item{\code{management_unit}: Management unit corresponding to each stand.}
\item{\code{represented_area}: Area represented by each stand (in hectares).}
}}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{meteo}{Meteorology data (see \code{\link{fordyn_spatial}}).}

\item{management_scenario}{A list defining the management scenario (see \code{\link{create_management_scenario}})}

\item{volume_function}{A function accepting a forest object as input and returning the wood volume (m3/ha) corresponding to each tree cohort.
If NULL, the default volume function is used (not recommended!).}

\item{local_control}{A list of local model control parameters (see \code{\link{defaultControl}}).}

\item{dates}{A \code{\link{Date}} object with the days of the period to be simulated. If \code{NULL}, then the whole period of \code{meteo} is used.}

\item{CO2ByYear}{A named numeric vector with years as names and atmospheric CO2 concentration (in ppm) as values. Used to specify annual changes in CO2 concentration along the simulation (as an alternative to specifying daily values in \code{meteo}).}

\item{summary_function}{An appropriate function to calculate summaries from an object of class 'fordyn' (e.g., \code{\link{summary.fordyn}}).}

\item{summary_arguments}{List with additional arguments for the summary function.}

\item{parallelize}{Boolean flag to try parallelization (will use all clusters minus one).}

\item{num_cores}{Integer with the number of cores to be used for parallel computation.}

\item{chunk_size}{Integer indicating the size of chunks to be sent to different processes (by default, the number of spatial elements divided by the number of cores).}

\item{progress}{Boolean flag to display progress information for simulations.}
}
\value{
An list of class 'fordyn_scenario' with the following elements:
\itemize{
\item{\code{sf}: An object of class 'sf' containing four elements:
\itemize{
\item{\code{geometry}: Spatial geometry.}
\item{\code{id}: Stand id, taken from the input.}
\item{\code{state}: A list of \code{\link{spwbInput}} or \code{\link{growthInput}} objects for each simulated stand, to be used in subsequent simulations (see \code{\link{update_landscape}}).}
\item{\code{forest}: A list of \code{\link{forest}} objects for each simulated stand (only in \code{fordynspatial}), to be used in subsequent simulations (see \code{\link{update_landscape}}).}
\item{\code{management_arguments}: A list of management arguments for each simulated stand (only in \code{fordynspatial}), to be used in subsequent simulations (see \code{\link{update_landscape}}).}
\item{\code{tree_table}: A list of data frames for each simulated stand, containing the living trees at each time step.}
\item{\code{shrub_table}: A list of data frames for each simulated stand, containing the living shrub at each time step.}
\item{\code{dead_tree_table}: A list of data frames for each simulated stand, containing the dead trees at each time step.}
\item{\code{dead_shrub_table}: A list of data frames for each simulated stand, containing the dead shrub at each time step.}
\item{\code{cut_tree_table}: A list of data frames for each simulated stand, containing the cut trees at each time step.}
\item{\code{cut_shrub_table}: A list of data frames for each simulated stand, containing the cut shrub at each time step.}
\item{\code{summary}: A list of model output summaries for each simulated stand (if \code{summary_function} was not \code{NULL}).}
}
}
\item{\code{volumes}: A data frame with extracted volumes (m3) by species and year. In demand-based scenarios target volumes are also included.}
}
}
\description{
Evaluates forest dynamics over a landscape including climate and management scenarios
}
\details{
Management is implemented using the \code{\link{defaultManagementFunction}} in medfate,
meaning that management parameters need to follow the structure of \code{\link{defaultManagementArguments}}
}
\examples{
\dontrun{
# Load example landscape data
data("examplepointslandscape")
  
# Transform example to 'sf' 
pts_sf <- sp_to_sf(examplepointslandscape)

# Load example meteo data frame from package meteoland
data("examplemeteo")
  
#Prepare a three-year meteorological data 
meteo_01_03 <- rbind(examplemeteo, examplemeteo, examplemeteo)
row.names(meteo_01_03) <- seq(as.Date("2001-01-01"), 
                              as.Date("2003-12-31"), by="day")
                         
# Load default medfate parameters
data("SpParamsMED")

# Creates scenario with one management unit and annual demand for P. nigra 
scen <- create_management_scenario(1, c("Pinus nigra" = 2300))

# Assign management unit to all stands
pts_sf$management_unit <- 1 

# Assume that each stand represents 1km2 = 100 ha
pts_sf$represented_area <- 100

# Launch simulation scenario
res <- fordyn_scenario(pts_sf, SpParamsMED, meteo = meteo_01_03, 
                       volume_function = NULL, management_scenario = scen,
                       parallelize = TRUE)
}

}
\seealso{
\code{\link{fordyn_spatial}}, \code{\link{create_management_scenario}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
