% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_forests.R
\name{impute_forests}
\alias{impute_forests}
\title{Forest imputation}
\usage{
impute_forests(
  x,
  sf_nfi,
  dem,
  forest_map,
  height_map = NULL,
  basal_area_map = NULL,
  var_class = NA,
  var_height = NA,
  var_basal_area = NA,
  max_distance_km = 50,
  replace_existing = FALSE,
  empty_forests = TRUE,
  correct_height_shrubs = FALSE,
  merge_trees = TRUE,
  merge_shrubs = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{An object of class \code{\link{sf}}. If it contains a column named 'land_cover_type', imputation
will be performed for locations whose land cover is "wildland". Otherwise, forest imputation is done for all locations.}

\item{sf_nfi}{An object of class \code{\link{sf}} with forest inventory data column 'forest'.}

\item{dem}{A digital elevation model (class \code{\link{rast}}) with meters as units}

\item{forest_map}{An object of class \code{\link{rast}} or \code{\link{vect}} with the forest class map}

\item{height_map}{An object of class \code{\link{rast}} or \code{\link{vect}} with the forest height map}

\item{basal_area_map}{An object of class \code{\link{rast}} or \code{\link{vect}} with the forest basal area map}

\item{var_class}{Variable name or index containing forest classes in 'forest_map'. If missing the first column is taken.}

\item{var_height}{Variable name or index containing forest height in 'height_map'. If missing the first column is taken.}

\item{var_basal_area}{Variable name or index containing forest height in 'basal_area_map'. If missing the first column is taken.}

\item{max_distance_km}{Maximum distance, in km, for forest inventory plot imputation.}

\item{replace_existing}{A logical flag to force the replacement of existing 'forest' objects, when present.}

\item{empty_forests}{A logical flag to place empty 'forest' objects for wildland cells not having a defined forest class.}

\item{correct_height_shrubs}{A logical flag to correct the height of shrubs when providing a vegetation height map (by default, only tree heights are corrected).}

\item{merge_trees}{A logical flag to simplify tree cohorts by merging tree records in DBH classes (see \code{\link{forest_mergeTrees}}).}

\item{merge_shrubs}{A logical flag to simplify shrub cohorts by merging shrub records in height classes (see \code{\link{forest_mergeShrubs}}).}

\item{verbose}{A logical flag to print console output.}
}
\value{
A modified object of class \code{\link{sf}} with column 'forest'.
}
\description{
Performs imputation of forest objects from a forest inventory using a forest map to match forest types and topography as covariates. Additional
rasters can be supplied to correct forest structure in the target locations.
}
\details{
The function performs a simplistic imputation of forest inventory plots on target locations provided that
they correspond to the same forest class, defined in the forest map, and are within some distance of the target location.
Among the multiple stands that can have the target forest class, the function chooses the one that has the most similar elevation
and position in the N-to-S slopes (i.e. the product of the cosine of aspect and slope). Both topographic
features are standardized to zero mean and unit standard deviation, to make their influence on the imputation
equal. This imputation method will be more or less successful depending on the resolution of forest classes and
the number of forest inventory plots available for each of them.

When \code{height_map} or \code{basal_area_map} are provided, the function performs structural corrections in additional to imputation.
When correcting for height, the function does not force average tree height of the stand in the target location to match the height map.
Rather, it uses the ratio of heights between the target location and the forest inventory plot used as imputation source, to correct the
tree heights of the target location. Tree diameters are corrected with the same factor, assuming that the diameter-height relationship
does not depend on tree height. This leads consequently in a change in basal area, in addition to changes in dominant tree height.

Correction for basal area operates on tree density. Analogously to the correction of heights, it does not take the basal area value of the map
for the target location, but the ratio of values between the target location and the forest inventory plot used as reference for imputation.
This ratio is used to correct tree density.
}
\examples{
\dontrun{
  # See vignette 'Landscape inputs'
}
}
\seealso{
\code{\link[=create_landscape]{create_landscape()}}, \code{\link[=add_soilgrids]{add_soilgrids()}}, \code{\link[medfate]{forest_mergeTrees}}
}
