% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medfateland-package.R
\docType{package}
\name{medfateland-package}
\alias{medfateland}
\alias{medfateland-package}
\title{medfateland: Mediterranean Landscape Forest Simulation}
\description{
Functions to simulate hydrology, forest functioning and dynamics over landscapes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emf-creaf.github.io/medfateland/index.html}
}

}
\author{
\strong{Maintainer}: Miquel De Cáceres \email{miquelcaceres@gmail.com}

Authors:
\itemize{
  \item Aitor Améztegui (\href{https://orcid.org/0000-0003-2006-1559}{ORCID})
  \item María González (\href{https://orcid.org/0000-0002-2227-8404}{ORCID})
}

Other contributors:
\itemize{
  \item Mario Beltrán [contributor]
}

}
\keyword{internal}
