% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medfateland-package.R
\docType{package}
\name{medfateland-package}
\alias{medfateland-package}
\alias{medfateland}
\title{medfateland: Mediterranean Landscape Forest Simulation}
\description{
Functions to simulate forest hydrology, forest function and dynamics over landscapes
}
\seealso{
Useful links: \itemize{ \item{
\url{https://emf-creaf.github.io/medfateland/index.html}} }
}
\author{
Maintainer: Miquel De Cáceres
\email{miquelcaceres@gmail.com}
\href{https://orcid.org/0000-0001-7132-2080}{ORCID}

Author: Aitor Améztegui
\href{https://orcid.org/0000-0003-2006-1559}{ORCID}

Author: María González
\href{https://orcid.org/0000-0002-2227-8404}{ORCID}

Author: Núria Aquilué

Contributor: Mario Beltrán

Contributor: Rodrigo Balaguer-Romano
\href{https://orcid.org/0000-0003-2808-6777}{ORCID}

Contributor: Roberto Molowny-Horas
\href{https://orcid.org/0000-0003-2626-6379}{ORCID}
}
\keyword{internal}
