% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DefineSpatialClasses.R
\docType{class}
\name{SpatialGridLandscape-class}
\alias{SpatialGridLandscape-class}
\title{Class \code{"SpatialGridLandscape"}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

An S4 class that represents a landscape configuration over a grid of coordinates.
}
\section{Slots}{

\describe{
\item{\code{grid}}{Object of class \code{"GridTopology"}.}

\item{\code{bbox}}{Object of class \code{"matrix"} with the boundary box.}

\item{\code{proj4string}}{Object of class \code{"CRS"} with the projection string.}

\item{\code{data}}{Object of class "data.frame" containing the elevation (in m), slope (in degrees) and aspect (in degrees from North) of every cell.}

\item{\code{lct}}{A character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').}

\item{\code{forestlist}}{Object of class \code{"list"} containing a set of \code{\link{forest}} objects.}

\item{\code{soillist}}{Object of class \code{"list"} containing a set of \code{\link{soil}} objects.}

\item{\code{xlist}}{Object of class \code{"list"} containing a set of spwb or growth model input objects.}
}}

\examples{
#Structure of the S4 object
showClass("SpatialGridLandscape")

}
\seealso{
\code{\link{forest}}, \code{\link{soil}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
