% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DefineSpatialClasses.R
\docType{class}
\name{SpatialPixelsLandscape-class}
\alias{SpatialPixelsLandscape-class}
\title{Class \code{"SpatialPixelsLandscape"}}
\arguments{
\item{data}{Object of class "data.frame" containing the elevation (in m), slope (in degrees) and aspect (in degrees from North) of every cell.}

\item{coords}{A matrix of pixel coordinates.}

\item{coords.nrs}{Inherited from \code{SpatialPointsDataFrame} but not used.}

\item{grid}{Object of class \code{GridTopology}.}

\item{grid.index}{Index of points in full grid.}

\item{bbox}{Object of class \code{"matrix"} with the boundary box.}

\item{proj4string}{Object of class \code{"CRS"} with the projection string.}

\item{lct}{A character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').}

\item{forestlist}{Object of class \code{"list"} containing a set of \code{\link{forest}} objects.}

\item{soillist}{Object of class \code{"list"} containing a set of \code{\link{soil}} objects.}

\item{xlist}{Object of class \code{"list"} containing a set of spwb or growth model input objects.}
}
\description{
`r lifecycle::badge("deprecated")`

An S4 class that represents a landscape configuration over an (incomplete) grid of coordinates.
}
\examples{
#Structure of the S4 object
showClass("SpatialPixelsLandscape")

}
\seealso{
\code{\link{spatialForestSummary}}, \code{\link{forest}}, \code{\link{soil}}, \code{\link{getLandscapeLayer}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
