% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_DefineSpatialClasses.R
\docType{class}
\name{SpatialPointsLandscape-class}
\alias{SpatialPointsLandscape-class}
\title{Class \code{"SpatialPointsLandscape"}}
\arguments{
\item{data}{Object of class "data.frame" containing the elevation (in m), slope (in degrees) and aspect (in degrees from North) of every cell.}

\item{coords}{Object of class \code{"matrix"} with spatial coordinates.}

\item{bbox}{Object of class \code{"matrix"} with the boundary box.}

\item{proj4string}{Object of class \code{"CRS"} with the projection string.}

\item{lct}{A character vector with the land cover type of each grid cell (values should be 'wildland', 'agriculture', 'rock', 'artificial' or 'water').}

\item{forestlist}{Object of class \code{"list"} containing a set of \code{\link{forest}} objects.}

\item{soillist}{Object of class \code{"list"} containing a set of \code{\link{soil}} objects.}

\item{xlist}{Object of class \code{"list"} containing a set of spwb or growth model input objects.}
}
\description{
An S4 class that represents a set of forest stands along with their spatial location
}
\examples{
showClass("SpatialPointsLandscape")

}
\seealso{
\code{\link{spatialForestSummary}}, \code{\link{forest}}, \code{\link{soil}}, \code{\link{getLandscapeLayer}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
